refactor: 패키지 구조 변경
- card 관련 패키지를 ui 하위로 이동
