+++ app/src/androidTest/java/woowacourse/payments/card/register/component/CardPasswordTextFieldTest.kt
@@ -12,6 +12,7 @@ import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
+import woowacourse.payments.ui.card.register.component.CardPasswordTextField
 
 @RunWith(AndroidJUnit4::class)
 class CardPasswordTextFieldTest {
