+++ app/src/main/java/woowacourse/payments/ui/card/component/PaymentCard.kt
@@ -1,11 +1,10 @@
-package woowacourse.payments.card.component
+package woowacourse.payments.ui.card.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
