+++ app/src/main/java/woowacourse/payments/ui/card/list/NoCardScreen.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.card.list
+package woowacourse.payments.ui.card.list
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
@@ -13,7 +13,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import woowacourse.payments.card.component.NewCard
+import woowacourse.payments.ui.card.component.NewCard
 
 @Composable
 fun NoCardScreen(onAddNewCardClick: () -> Unit) {
